pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "imagekitio>=5.4.0",
    "langchain>=1.2.15",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.2",
]
